Traktor: Implemented a "facade" component which can turn child entities on and off. Also refactored RefSet to use SmallSet instead of stl set.
[git-p4: depot-paths = "//depot/traktor/": change = 24007]

+++ code/Render/Editor/Shader/ShaderViewer.cpp
@@ -432,10 +432,10 @@ void ShaderViewer::jobReflect(Ref< ShaderGraph > shaderGraph, Ref< const IProgra
 
 					std::vector< Guid >::iterator it = std::find(textureIds.begin(), textureIds.end(), textureGuid);
 					if (it != textureIds.end())
-						textureIndex = std::distance(textureIds.begin(), it);
+						textureIndex = (int32_t)std::distance(textureIds.begin(), it);
 					else
 					{
-						textureIndex = int32_t(textureIds.size());
+						textureIndex = (int32_t)textureIds.size();
 						textureIds.push_back(textureGuid);
 					}
